removes bit_strea_parser class, combines common stuff to roc_video_parser file
